fixed linting on schema
